+++ crates/miden-block-prover/src/lib.rs
@@ -3,6 +3,3 @@ pub use errors::ProvenBlockError;
 
 mod local_block_prover;
 pub use local_block_prover::LocalBlockProver;
-
-#[cfg(test)]
-mod tests;