+++ crates/miden-testing/src/mock_chain/mod.rs
@@ -1,0 +1,8 @@
+mod account;
+mod auth;
+mod chain;
+mod fungible_faucet;
+
+pub use auth::Auth;
+pub use chain::{AccountState, MockChain};
+pub use fungible_faucet::MockFungibleFaucet;