+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Rename `ChainMmr` to `PartialBlockchain` (#1305).
 - Add safe `PartialBlockchain` constructor (#1308).
 - Fix error when creating accounts with empty storage (#1307).
+- [BREAKING] Move `MockChain` and `TransactionContext` to new `miden-testing` crate (#1309).
 
 ## 0.8.2 (2025-04-18) - `miden-proving-service` crate only